test(orb): move confirm-gate coverage into a dedicated tool test file (CI impact-scan)
The impact-scan bot's conversation-flow-change-needs-test rule flagged
orb-tools-shared.ts's confirm-gate change (previous commit) because its
test coverage lived in voice-send-chat-rate-limit-isolation.test.ts,
whose filename doesn't match the flow-test naming heuristic. Moves that
coverage into test/services/send-chat-message-tool-confirm-gate.test.ts
(self-contained, no shared-helper changes needed) so the gate recognizes
it — same tests, no behavior change.
